REL: v0.2.0 (#1)
* ENH: Get k8s deployment working, add YTT manifests

* CONT: More of the last commit that I forgot to add

* ENH: Fix default config

* ENH: Fix node management

* FIX: Resolve https connection and kapp ordering issues

* MAINT: Update gitignore

* ENH: Fix various bugs, resolve UI display issue

* DOC: Update README

* MAINT: Update copyright on docs

* ENH: CLI updates

* ENH: Server updates

* MAINT: Misc updates

* MAINT: Rename ui dir

* FIX: Resolve auto-execute issue

* ENH: Add ability to change auto-execute from UI

* ENH: Allow cascade delete from UI

* ENH: Add local run support

* ENH: Fix Bulwark connection issues, update packages

+++ src/server/container/container.go
@@ -4,10 +4,11 @@ import (
 	"fmt"
 	"os"
 	"os/exec"
-	"scaffold/server/logger"
 	"scaffold/server/run"
 	"strings"
 	"time"
+
+	logger "github.com/jfcarter2358/go-logger"
 )
 
 var LastRun []string
@@ -35,17 +36,21 @@ func PruneContainers() {
 
 			parts := strings.Split(toDestroy, ".")
 
-			containerName := fmt.Sprintf("%s-%s", parts[0], parts[1])
-			if containerName != CurrentName {
-				logs, err := exec.Command("/bin/sh", "-c", fmt.Sprintf("podman rm %s", containerName)).CombinedOutput()
-				if err != nil {
-					logger.Errorf("", "Remove error string: %s", err.Error())
+			if len(parts) == 2 {
+				containerName := fmt.Sprintf("%s-%s", parts[0], parts[1])
+				if containerName != CurrentName {
+					logs, err := exec.Command("/bin/sh", "-c", fmt.Sprintf("podman rm %s", containerName)).CombinedOutput()
+					if err != nil {
+						logger.Errorf("", "Remove error string: %s", err.Error())
+					}
+					logger.Debugf("", "Remove output: %s", logs)
+					runDir := fmt.Sprintf("/tmp/run/%s/%s/%s", parts[0], parts[1], parts[2])
+					if err := os.RemoveAll(runDir); err != nil {
+						logger.Errorf("", "Delete error string: %s", err.Error())
+					}
 				}
-				logger.Debugf("", "Remove output: %s", logs)
-				runDir := fmt.Sprintf("/tmp/run/%s/%s/%s", parts[0], parts[1], parts[2])
-				if err := os.RemoveAll(runDir); err != nil {
-					logger.Errorf("", "Delete error string: %s", err.Error())
-				}
+			} else {
+				logger.Debugf("", "Got %v when expected length 2", parts)
 			}
 		}
 		time.Sleep(250 * time.Millisecond)
